pyproject.toml
[tool.poetry]
name = "rag-ai-stack-v02"
version = "0.1.0"
description = "Simple but modular and extensible RAG AI stack to enable you to locally chat with your own documents and other sources."
authors = ["Jonas A. Schmidt <jonas.a.schmidt@gmail.com>"]

# Deaktiviere Paket-Modus (nur für deps, nicht als installierbares Projekt)
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
llama-index = "^0.10.0"
chainlit = "^1.0.0"
watchdog = "^4.0.0"
python-dotenv = "^1.0.0"
pypdf = "^4.3.1"

[tool.poetry.group.dev.dependencies]
black = "^24.0"
pytest = "^7.0"
pre-commit = "^4.2.0"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"